Move from_parquet loading to (spacecraft) trajectory directly and remove trajectory_data
