+++ src/od/simulator/scheduler.rs
@@ -26,12 +26,11 @@ use serde_derive::Serialize;
 use std::fmt::Debug;
 use typed_builder::TypedBuilder;
 
-#[cfg(feature = "python")]
-use pyo3::prelude::*;
+
 
 /// Defines the handoff from a current ground station to the next one that is visible to prevent overlapping of measurements
 #[derive(Copy, Clone, Debug, Deserialize, PartialEq, Serialize)]
-#[cfg_attr(feature = "python", pyclass)]
+
 #[cfg_attr(feature = "python", pyo3(module = "nyx_space.orbit_determination"))]
 pub enum Handoff {
     /// If a new station is in visibility of the spacecraft, the "Eager" station will immediately stop tracking and switch over (default)
@@ -51,7 +50,7 @@ impl Default for Handoff {
 /// A scheduler allows building a scheduling of spaceraft tracking for a set of ground stations.
 #[derive(Copy, Clone, Debug, Default, Deserialize, PartialEq, Serialize, TypedBuilder)]
 #[builder(doc)]
-#[cfg_attr(feature = "python", pyclass)]
+
 #[cfg_attr(feature = "python", pyo3(module = "nyx_space.orbit_determination"))]
 pub struct Scheduler {
     /// Handoff strategy if two trackers see the vehicle at the same time
