--- tests/python/.gitignore
@@ -1,3 +1,0 @@
-__pycache__
-private
-mod.rs
