--- tests/python/README.md
@@ -1,3 +1,0 @@
-# Python tests
-
-The Python tests only tests the functionality and usability. All other tests are done in Rust with the `cargo test` command. This is why the Python tests are not exhaustive.
